refactor block manager module
