+++ reductstore/src/storage/block_manager.rs
@@ -6,7 +6,12 @@ mod block_cache;
 pub(in crate::storage) mod block_index;
 pub(in crate::storage) mod compress;
 pub(in crate::storage) mod decompress_cache;
+mod lifecycle;
+mod lookup;
+mod paths;
+mod persistence;
 mod read_only;
+mod records;
 pub(in crate::storage) mod wal;
 
 use crate::backend::BackendType;
@@ -104,864 +109,6 @@ impl BlockManager {
             usage_counters,
             last_replica_sync: Instant::now(),
         })
-    }
-
-    pub async fn save_cache_on_disk(&mut self) -> Result<(), ReductError> {
-        let blocks = self.block_cache.write_values();
-        for block in blocks.iter() {
-            {
-                let block_id = block.read().await?.block_id();
-                self.sync_data_block(block_id).await?;
-            }
-            self.save_meta_on_disk(block.clone()).await?;
-        }
-
-        Ok(())
-    }
-
-    pub async fn save_cache_metadata_on_disk(&mut self) -> Result<(), ReductError> {
-        let blocks_with_wal = self.wal.list().await.unwrap_or_default();
-        if blocks_with_wal.is_empty() {
-            return Ok(());
-        }
-
-        let blocks = self.block_cache.write_values();
-        for block in blocks {
-            let block_id = block.read().await?.block_id();
-            if blocks_with_wal.contains(&block_id) {
-                self.save_meta_on_disk(block).await?;
-            }
-        }
-
-        Ok(())
-    }
-
-    pub async fn find_block(&mut self, start: u64) -> Result<BlockRef, ReductError> {
-        let id = self.find_block_id(start)?;
-        self.load_block(id).await
-    }
-
-    fn find_block_id(&self, start: u64) -> Result<u64, ReductError> {
-        let active_tree = self.block_index.active_tree();
-        let start_block_id = active_tree.range(start..).next();
-        if start_block_id.is_some() && start >= *start_block_id.unwrap() {
-            Ok(*start_block_id.unwrap())
-        } else if let Some(block_id) = active_tree.range(..start).rev().next() {
-            Ok(*block_id)
-        } else {
-            Err(ReductError::not_found(&format!(
-                "Record {} not found in entry {}/{}",
-                start, self.bucket, self.entry
-            )))
-        }
-    }
-
-    pub fn find_cached_block(&self, start: u64) -> Option<BlockRef> {
-        let id = self.find_block_id(start).ok()?;
-
-        self.block_cache.get_read(&id)
-    }
-
-    pub async fn load_block(&mut self, block_id: u64) -> Result<BlockRef, ReductError> {
-        // first check if we have the block in write cache
-        let mut cached_block = self.block_cache.get_read(&block_id);
-        if cached_block.is_none() {
-            let path = self.resolve_desc_path(block_id).await?;
-            let buf = match FILE_CACHE.read(&path, SeekFrom::Start(0)).await {
-                Ok(mut file) => {
-                    let mut buf = vec![];
-                    file.read_to_end(&mut buf)?;
-                    buf
-                }
-                Err(err) => {
-                    // Re-check existence to distinguish a transient TOCTOU race
-                    // (descriptor removed after the first check) from a real read failure.
-                    if self.cfg.role == InstanceRole::Replica
-                        && !FILE_CACHE.try_exists(&path).await?
-                    {
-                        self.block_index.remove_block(block_id);
-                        return Err(too_early!(
-                            "Block descriptor {:?} can't be read on replica yet: {}. Reload index and retry",
-                            path,
-                            err
-                        ));
-                    }
-
-                    // here we can't read the block descriptor, it might be corrupted or not exist
-                    let err_msg = format!("Block descriptor {:?} can't be read: {}", path, err);
-                    error!("{}", &err_msg);
-                    self.mark_block_corrupted(block_id).await?;
-                    return Err(internal_server_error!(&err_msg));
-                }
-            };
-
-            // calculate crc of the block descriptor
-            let mut crc = Digest::new();
-            crc.write(&buf);
-
-            if let Some(block) = self.block_index.get_block(block_id) {
-                if let Some(block_crc) = block.crc64 {
-                    // we check crc if the crc is stored in the index for backward compatibility
-                    if block_crc != crc.sum64() {
-                        if self.cfg.role == InstanceRole::Replica {
-                            warn!(
-                                "Block descriptor {:?} CRC mismatch on replica: index CRC {} mismatch with calculated CRC {}. Treat as transient and reload index",
-                                path,
-                                block_crc,
-                                crc.sum64()
-                            );
-                            self.invalidate_replica_block_cache(block_id).await?;
-                            self.block_index.remove_block(block_id);
-                            return Err(too_early!(
-                                "Block descriptor {:?} CRC mismatch on replica. Reload index and retry",
-                                path
-                            ));
-                        }
-
-                        match BlockProto::decode(Bytes::from(buf.clone())) {
-                            Ok(decoded) => {
-                                let desc_version = decoded.version.unwrap_or(0);
-                                let index_version = block.version.unwrap_or(0);
-
-                                if desc_version > index_version {
-                                    warn!(
-                                        "Block descriptor {:?} is newer than index (v{} > v{}). Updating index",
-                                        path, desc_version, index_version
-                                    );
-
-                                    self.block_index
-                                        .insert_or_update_with_crc(decoded.clone(), crc.sum64());
-                                    self.block_index.save().await?;
-                                } else {
-                                    error!(
-                                        "Block descriptor {:?} is corrupted: index CRC {} mismatch with calculated CRC {}, version {} <= {}",
-                                        path,
-                                        block_crc,
-                                        crc.sum64(),
-                                        desc_version,
-                                        index_version
-                                    );
-
-                                    self.mark_block_corrupted(block_id).await?;
-                                    return Err(internal_server_error!(
-                                        "Block descriptor {:?} is corrupted",
-                                        path
-                                    ));
-                                }
-                            }
-                            Err(err) => {
-                                error!(
-                                    "Block descriptor {:?} is corrupted: index CRC {} mismatch with calculated CRC {} and descriptor can't be decoded: {}",
-                                    path,
-                                    block_crc,
-                                    crc.sum64(),
-                                    err
-                                );
-                                self.mark_block_corrupted(block_id).await?;
-                                return Err(internal_server_error!(
-                                    "Block descriptor {:?} is corrupted",
-                                    path
-                                ));
-                            }
-                        }
-                    }
-                }
-            } else {
-                return Err(internal_server_error!(
-                    "Block descriptor {:?} is not in the index",
-                    path
-                ));
-            }
-
-            // parse the block descriptor
-            let mut block_from_disk = match BlockProto::decode(Bytes::from(buf)) {
-                Ok(block) => block,
-                Err(e) => {
-                    self.mark_block_corrupted(block_id).await?;
-                    return Err(internal_server_error!(
-                        "Failed to decode block descriptor {:?}: {}",
-                        path,
-                        e
-                    ));
-                }
-            };
-
-            if block_from_disk.begin_time.is_none() {
-                warn!(
-                    "Block descriptor {:?} has no begin time. It might be recovered from the WAL",
-                    path
-                );
-                block_from_disk.begin_time = Some(us_to_ts(&block_id));
-            }
-
-            cached_block = Some(Arc::new(AsyncRwLock::new(block_from_disk.into())));
-        }
-
-        let cached_block = cached_block.unwrap();
-        self.block_cache.insert_read(block_id, cached_block.clone());
-        Ok(cached_block)
-    }
-
-    pub async fn save_block(&mut self, block: BlockRef) -> Result<(), ReductError> {
-        let id = block.read().await?.block_id();
-        for (_, block) in self.block_cache.insert_write(id, block.clone()) {
-            self.save_meta_on_disk(block).await?;
-        }
-
-        Ok(())
-    }
-
-    pub async fn start_new_block(
-        &mut self,
-        block_id: u64,
-        max_block_size: u64,
-    ) -> Result<BlockRef, ReductError> {
-        let block = Block::new(block_id);
-
-        // create a block with data
-        {
-            let mut file = FILE_CACHE
-                .write_or_create(&self.path_to_data(block_id), SeekFrom::Start(0))
-                .await?;
-
-            if self.cfg.backend_config.backend_type == BackendType::Filesystem {
-                // Pre-allocation for remote storage is inefficient
-                // because we synchronize full size even if block is empty
-                file.set_len(max_block_size)?;
-            }
-        }
-
-        self.block_index.insert_or_update(block.clone());
-
-        let block_ref = Arc::new(AsyncRwLock::new(block));
-        self.save_block(block_ref.clone()).await?;
-        Ok(block_ref)
-    }
-
-    /// Finish writing a record to a block.
-    ///
-    /// This method will shrink the block data file to the record size and sync the descriptor and data files.
-    ///
-    /// # Arguments
-    ///
-    /// * `block` - Block to finish writing to.
-    ///
-    /// # Errors
-    ///
-    /// * `ReductError` - If file system operation failed.
-    pub async fn finish_block(&mut self, block: BlockRef) -> Result<(), ReductError> {
-        let (block_id, block_size) = {
-            let block = block.read().await?;
-            (block.block_id(), block.size())
-        };
-
-        let data_path = self.path_to_data(block_id);
-        let desc_path = self.path_to_desc(block_id);
-        let index_path = self.path.join(BLOCK_INDEX_FILE);
-
-        {
-            // resize imminently for better testing.
-            let mut data_block = FILE_CACHE
-                .write_or_create(&data_path, SeekFrom::Current(0))
-                .await?;
-            data_block.set_len(block_size)?;
-        }
-
-        self.save_meta_on_disk(block.clone()).await?;
-
-        let sync_block = async move {
-            /* sync descriptor and data */
-            {
-                let mut data_block = FILE_CACHE
-                    .write_or_create(&data_path, SeekFrom::Current(0))
-                    .await?;
-                data_block.sync_all().await?;
-            }
-
-            {
-                let mut descr_block = FILE_CACHE
-                    .write_or_create(&desc_path, SeekFrom::Current(0))
-                    .await?;
-                descr_block.sync_all().await?;
-            }
-
-            {
-                let mut descr_block = FILE_CACHE
-                    .write_or_create(&desc_path, SeekFrom::Current(0))
-                    .await?;
-                descr_block.sync_all().await?;
-            }
-
-            {
-                let mut index_file = FILE_CACHE
-                    .write_or_create(&index_path, SeekFrom::Current(0))
-                    .await?;
-                index_file.sync_all().await?;
-            }
-
-            Ok::<(), ReductError>(())
-        };
-
-        tokio::spawn(async move {
-            // spawn to avoid blocking entry
-            if let Err(err) = sync_block.await {
-                error!("{}", err)
-            }
-        });
-
-        Ok(())
-    }
-
-    /// Remove a block from file system.
-    ///
-    /// This method will sync the block descriptor and data files, remove the block from the cache and index.
-    ///
-    /// # Arguments
-    ///
-    /// * `block_id` - ID of the block to remove.
-    ///
-    /// # Errors
-    ///
-    /// * `ReductError` - If the block is still in use or file system operation failed.
-    pub async fn remove_block(&mut self, block_id: u64) -> Result<(), ReductError> {
-        self.wal.append(block_id, WalEntry::RemoveBlock).await?;
-
-        let data_block_path = self.path_to_data(block_id);
-        if FILE_CACHE.try_exists(&data_block_path).await? {
-            // it can be still in WAL only
-            FILE_CACHE.remove(&data_block_path).await?;
-        }
-
-        let desc_block_path = self.path_to_desc(block_id);
-        if FILE_CACHE.try_exists(&desc_block_path).await? {
-            // it can be still in WAL only
-            FILE_CACHE.remove(&desc_block_path).await?;
-        }
-
-        self.block_index.remove_block(block_id);
-        self.block_index.save().await?;
-
-        self.block_cache.remove(&block_id);
-
-        self.wal.remove(block_id).await?;
-        Ok(())
-    }
-
-    pub async fn mark_block_corrupted(&mut self, block_id: u64) -> Result<(), ReductError> {
-        if self.cfg.role == InstanceRole::Replica {
-            return Ok(());
-        }
-
-        self.block_index.mark_corrupted(block_id);
-        self.block_cache.remove(&block_id);
-
-        let path = self.path_to_desc(block_id);
-        let descriptor = if let Ok(mut file) = FILE_CACHE.read(&path, SeekFrom::Start(0)).await {
-            let mut buf = Vec::new();
-            if file.read_to_end(&mut buf).is_ok() {
-                BlockProto::decode(Bytes::from(buf)).ok()
-            } else {
-                None
-            }
-        } else {
-            None
-        };
-
-        if let Some(mut proto) = descriptor {
-            proto.corrupted = Some(true);
-            let new_buf = proto.encode_to_vec();
-            if let Ok(mut writer) = FILE_CACHE.write_or_create(&path, SeekFrom::Start(0)).await {
-                let _ = writer.set_len(new_buf.len() as u64);
-                let _ = writer.write_all(&new_buf);
-                let _ = writer.flush_local().await;
-            }
-        }
-
-        self.block_index.save().await?;
-
-        error!(
-            "Marked block {}/{}/{} as corrupted",
-            self.bucket, self.entry, block_id
-        );
-        Ok(())
-    }
-
-    #[cfg(test)]
-    pub fn is_block_corrupted(&self, block_id: u64) -> bool {
-        self.block_index.is_corrupted(block_id)
-    }
-
-    /// Check if a block exists on disk.
-    pub async fn exist(&self, block_id: u64) -> Result<bool, ReductError> {
-        let path = self.path_to_desc(block_id);
-        Ok(FILE_CACHE.try_exists(&path).await?)
-    }
-
-    /// Update records in a block and save it on disk.
-    ///
-    /// # Arguments
-    ///
-    /// * `block_id` - Block to update records in.
-    /// * `records` - Records to update.
-    ///
-    ///
-    /// # Panics
-    ///
-    /// * If the record is not in the block.
-    ///
-    /// # Returns
-    ///
-    /// * `Ok(())` - If the records were updated successfully.
-    ///
-    /// # Errors
-    ///
-    /// * `ReductError` - If failed to append to WAL or save the block on disk.
-    pub async fn update_records(
-        &mut self,
-        block_id: u64,
-        records: Vec<Record>,
-    ) -> Result<(), ReductError> {
-        self.decompress_block(block_id).await?;
-        let block_ref = self.load_block(block_id).await?;
-
-        // First, append all WAL entries
-        for record in records.iter() {
-            self.wal
-                .append(block_id, WalEntry::UpdateRecord(record.clone()))
-                .await?;
-        }
-
-        // Then, update the block in-memory (no await needed here)
-        {
-            let mut block = block_ref.write().await?;
-            for record in records.into_iter() {
-                block.insert_or_update_record(record);
-            }
-        }
-
-        self.save_block(block_ref).await
-    }
-
-    /// Remove records from a block and save it on disk.
-    ///
-    /// The method will create a temporary block file and copy the retained records to it.
-    /// In the end, the original block file will be replaced with the temporary one and the block descriptor will be updated.
-    ///
-    /// # Arguments
-    ///
-    /// * `block_id` - Block to remove records from.
-    /// * `records` - Record timestamps to remove.
-    ///
-    /// # Returns
-    ///
-    /// * `Ok(())` - If the records were removed successfully.
-    pub async fn remove_records(
-        &mut self,
-        block_id: u64,
-        records: Vec<u64>,
-    ) -> Result<(), ReductError> {
-        self.decompress_block(block_id).await?;
-        let block_ref = self.load_block(block_id).await?;
-
-        {
-            let mut block = block_ref.write().await?;
-            for record_time in records {
-                block.remove_record(record_time);
-                self.wal
-                    .append(block_id, WalEntry::RemoveRecord(record_time))
-                    .await?;
-            }
-
-            // if the block is empty, remove it
-            if block.record_count() == 0 {
-                let block_id = block.block_id();
-                drop(block);
-                return self.remove_block(block_id).await;
-            }
-        }
-
-        // Collect record info needed for copying
-        let (temp_block_path, record_info, src_block_path) = {
-            let block = block_ref.read().await?;
-            let temp_block_path = self.path.join(format!("{}.blk.tmp", block.block_id()));
-            let src_block_path = self.path_to_data(block.block_id());
-
-            // Collect record positions
-            let record_info: Vec<(u64, u64, u64)> = block
-                .record_index()
-                .values()
-                .map(|record| {
-                    let record_time = ts_to_us(&record.timestamp.unwrap());
-                    (record_time, record.begin, record.end - record.begin)
-                })
-                .collect();
-
-            (temp_block_path, record_info, src_block_path)
-        };
-
-        // create a temporary block file outside of the cache to avoid unnecessary evictions
-        let mut temp_block = OpenOptions::new()
-            .create(true)
-            .write(true)
-            .read(true)
-            .open(&temp_block_path)
-            .map_err(|e| {
-                internal_server_error!(
-                    "Failed to create temporary block file {:?}: {}",
-                    temp_block_path,
-                    e
-                )
-            })?;
-
-        // Copy records to temp file and track new positions
-        let mut new_positions: Vec<(u64, u64, u64)> = Vec::new();
-        let mut total_offset = 0u64;
-
-        for (record_time, begin, record_size) in record_info {
-            let mut read_bytes = 0;
-            while read_bytes < record_size {
-                let (buf, read) =
-                    read_in_chunks(&src_block_path, begin, record_size, read_bytes).await?;
-
-                read_bytes += read as u64;
-                temp_block.write_all(&buf)?;
-            }
-
-            let new_begin = total_offset;
-            total_offset += read_bytes;
-            let new_end = total_offset;
-
-            new_positions.push((record_time, new_begin, new_end));
-
-            trace!(
-                "Record {}/{} retained with new position begin={}, end={}",
-                block_id,
-                record_time,
-                new_begin,
-                new_end
-            );
-        }
-
-        // Update record positions in the block
-        {
-            let mut block = block_ref.write().await?;
-            for (record_time, new_begin, new_end) in new_positions {
-                if let Some(record) = block.record_index_mut().get_mut(&record_time) {
-                    record.begin = new_begin;
-                    record.end = new_end;
-                }
-            }
-
-            debug!(
-                "New block {:?} is created with {} records",
-                temp_block_path,
-                block.record_count()
-            );
-        }
-
-        // Replace the old block file with the new one
-        let block_path = {
-            let block = block_ref.read().await?;
-            self.path_to_data(block.block_id())
-        };
-
-        tokio::fs::remove_file(&block_path).await?;
-        tokio::fs::rename(&temp_block_path, &block_path).await?;
-
-        FILE_CACHE.discard_recursive(&block_path).await?;
-        let mut block_file = FILE_CACHE
-            .write_or_create(&block_path, SeekFrom::Start(0))
-            .await?;
-        block_file.sync_all().await?;
-
-        self.save_meta_on_disk(block_ref).await
-    }
-
-    /// Begin writing a record to a block.
-    ///
-    /// # Arguments
-    ///
-    /// * `block` - Block to write to.
-    /// * `record_timestamp` - Timestamp of the record to write.
-    ///
-    /// # Returns
-    ///
-    /// * `Ok(file)` - File to write to.
-    pub fn begin_write_record(
-        &self,
-        block: &Block,
-        record_timestamp: u64,
-    ) -> Result<(PathBuf, u64), ReductError> {
-        let path = self.path_to_data(block.block_id());
-        let offset = block.get_record(record_timestamp).unwrap().begin;
-        Ok((path, offset))
-    }
-
-    /// Finish writing a record to a block.
-    ///
-    /// This method will update the record state, add WAL entry and save the block on disk if needed.
-    ///
-    /// # Arguments
-    ///
-    /// * `block_id` - ID of the block to write to.
-    /// * `state` - State of the record.
-    ///
-    /// # Errors
-    ///
-    /// * `ReductError` - If file system operation failed.
-    pub async fn finish_write_record(
-        &mut self,
-        block_id: u64,
-        state: record::State,
-        record_timestamp: u64,
-    ) -> Result<(), ReductError> {
-        // check if the block is still in cache
-        let block_ref = if let Some(block_ref) = self.block_cache.get_write(&block_id) {
-            let block = block_ref.read().await?;
-            if block.block_id() == block_id {
-                block_ref.clone()
-            } else {
-                self.load_block(block_id).await?
-            }
-        } else {
-            self.load_block(block_id).await?
-        };
-
-        // Get the record for WAL entry before modifying
-        let wal_record = {
-            let mut block = block_ref.write().await?;
-            block.change_record_state(record_timestamp, i32::from(state))?;
-            block.get_record(record_timestamp).unwrap().clone()
-        };
-
-        // write to WAL (no guard held)
-        self.wal
-            .append(block_id, WalEntry::WriteRecord(wal_record))
-            .await?;
-
-        self.save_block(block_ref).await?;
-
-        debug!(
-            "Finished writing record {} to block {}/{}/{}.meta with state {:?}",
-            record_timestamp, self.bucket, self.entry, block_id, state
-        );
-
-        Ok(())
-    }
-
-    async fn sync_data_block(&self, block_id: u64) -> Result<(), ReductError> {
-        if self.cfg.role == InstanceRole::Replica {
-            return Ok(());
-        }
-
-        let path = self.path_to_data(block_id);
-        if !FILE_CACHE.try_exists(&path).await? {
-            return Ok(());
-        }
-
-        let mut data_block = FILE_CACHE
-            .write_or_create(&path, SeekFrom::Current(0))
-            .await?;
-        data_block.sync_all().await?;
-        Ok(())
-    }
-
-    /// Begin reading a record from a block.
-    ///
-    /// # Arguments
-    ///
-    /// * `block` - Block to read from.
-    /// * `record_timestamp` - Timestamp of the record to read.
-    ///
-    /// # Returns
-    ///
-    /// * `Ok(file, offset)` - File to read from and offset to start reading.
-    ///
-    /// # Errors
-    ///
-    /// * `ReductError` - If file system operation failed.
-    pub(crate) async fn begin_read_record(
-        &self,
-        block: &Block,
-        record_timestamp: u64,
-    ) -> Result<(PathBuf, u64), ReductError> {
-        let path = self.resolve_data_path(block.block_id()).await?;
-        let offset = block.get_record(record_timestamp).unwrap().begin;
-        Ok((path, offset))
-    }
-
-    async fn resolve_desc_path(&self, block_id: u64) -> Result<PathBuf, ReductError> {
-        if self.block_is_compressed(block_id) {
-            let compressed_desc_path = self.path_to_compressed_desc(block_id);
-            if FILE_CACHE.try_exists(&compressed_desc_path).await? {
-                self.decompress_cache
-                    .get_or_decompress(
-                        &self.path,
-                        block_id,
-                        DecompressedFileType::Descriptor,
-                        &compressed_desc_path,
-                    )
-                    .await
-            } else {
-                Ok(self.path_to_desc(block_id))
-            }
-        } else {
-            Ok(self.path_to_desc(block_id))
-        }
-    }
-
-    async fn resolve_data_path(&self, block_id: u64) -> Result<PathBuf, ReductError> {
-        if self.block_is_compressed(block_id) {
-            self.decompress_cache
-                .get_or_decompress(
-                    &self.path,
-                    block_id,
-                    DecompressedFileType::Data,
-                    &self.path_to_compressed_data(block_id),
-                )
-                .await
-        } else {
-            Ok(self.path_to_data(block_id))
-        }
-    }
-
-    pub(in crate::storage) fn block_is_compressed(&self, block_id: u64) -> bool {
-        self.block_index
-            .get_block(block_id)
-            .and_then(|block| block.compression)
-            .unwrap_or(i32::from(CompressionAlgorithm::None))
-            != i32::from(CompressionAlgorithm::None)
-    }
-
-    pub fn index_mut(&mut self) -> &mut BlockIndex {
-        &mut self.block_index
-    }
-
-    pub fn index(&self) -> &BlockIndex {
-        &self.block_index
-    }
-
-    pub(in crate::storage) fn usage_counters(&self) -> &UsageCounters {
-        &self.usage_counters
-    }
-
-    pub(in crate::storage) fn usage_counters_arc(&self) -> Arc<UsageCounters> {
-        Arc::clone(&self.usage_counters)
-    }
-
-    pub(in crate::storage) fn is_replica(&self) -> bool {
-        self.cfg.role == InstanceRole::Replica
-    }
-
-    async fn invalidate_replica_block_cache(&self, block_id: u64) -> Result<(), ReductError> {
-        FILE_CACHE
-            .invalidate_local_cache_file(&self.path_to_desc(block_id))
-            .await?;
-        FILE_CACHE
-            .invalidate_local_cache_file(&self.path_to_data(block_id))
-            .await?;
-        Ok(())
-    }
-
-    pub(in crate::storage) fn is_block_in_write_cache(&self, block_id: u64) -> bool {
-        self.block_cache.get_write(&block_id).is_some()
-    }
-
-    pub fn bucket_name(&self) -> &String {
-        &self.bucket
-    }
-
-    pub fn entry_name(&self) -> &String {
-        &self.entry
-    }
-
-    pub fn path(&self) -> &PathBuf {
-        &self.path
-    }
-
-    #[cfg(test)]
-    pub(crate) fn clear_cache_for_test(&self) {
-        self.block_cache.clear();
-    }
-
-    fn path_to_desc(&self, block_id: u64) -> PathBuf {
-        self.path
-            .join(format!("{}{}", block_id, DESCRIPTOR_FILE_EXT))
-    }
-
-    fn path_to_data(&self, block_id: u64) -> PathBuf {
-        self.path.join(format!("{}{}", block_id, DATA_FILE_EXT))
-    }
-
-    fn path_to_compressed_desc(&self, block_id: u64) -> PathBuf {
-        self.path
-            .join(format!("{}{}", block_id, COMPRESSED_DESCRIPTOR_FILE_EXT))
-    }
-
-    fn path_to_compressed_data(&self, block_id: u64) -> PathBuf {
-        self.path
-            .join(format!("{}{}", block_id, COMPRESSED_DATA_FILE_EXT))
-    }
-
-    // Method save descriptor and update index
-    // Note: it calls local sync but not sync_all to avoid blocking entry during synchronization with remote backend
-    // the blocks must be synced with backed from FILE_CACHE sync loop
-    async fn save_meta_on_disk(&mut self, block_ref: BlockRef) -> Result<(), ReductError> {
-        // Take a snapshot under a short-lived write lock to avoid blocking readers
-        let (block_id, block_snapshot) = {
-            let block = block_ref.read().await?;
-            (block.block_id(), block.to_owned())
-        };
-
-        debug!(
-            "Saving block {}/{}/{} on disk and updating index",
-            self.bucket, self.entry, block_id
-        );
-
-        let path = self.path_to_desc(block_id);
-        let mut buf = BytesMut::new();
-
-        let mut proto = BlockProto::from(block_snapshot);
-        let version = self
-            .block_index
-            .get_block(block_id)
-            .and_then(|block| block.version)
-            .unwrap_or(0)
-            + 1;
-        proto.version = Some(version);
-        proto.corrupted = None;
-        proto.encode(&mut buf).map_err(|e| {
-            internal_server_error!("Failed to encode block descriptor {:?}: {}", path, e)
-        })?;
-        let len = buf.len() as u64;
-
-        trace!("Writing block descriptor {:?}", path);
-
-        if self.cfg.role != InstanceRole::Replica {
-            let mut lock = FILE_CACHE
-                .write_or_create(&path, SeekFrom::Start(0))
-                .await?;
-            lock.set_len(len)?;
-            lock.write_all(&buf)?;
-            lock.flush_local().await?; // fix https://github.com/reductstore/reductstore/issues/642
-        }
-
-        trace!("Updating block index");
-        // update index with block crc
-        let mut crc = Digest::new();
-        crc.write(&buf);
-        proto.metadata_size = len; // update metadata size because it changed
-        self.block_index
-            .insert_or_update_with_crc(proto, crc.sum64());
-
-        if self.cfg.role != InstanceRole::Replica {
-            self.block_index.save().await?;
-
-            trace!("Block {}/{}/{} saved", self.bucket, self.entry, block_id);
-            // clean WAL
-            self.wal.remove(block_id).await?;
-        }
-        Ok(())
     }
 }
